Fix nodejs lambda asset hash
The asset hash was always the same. The issue is that we
create a stub archive first and cdk fingerprints that archive.
The archive is always empty. So no matter how many lambdas
you create or whether or not the code changes it would
retain the same signature. The lambda bundle now uses the
source as the assetHash if none is provided.

This fix required a patch to fs.realpathSync which allows
real path to return "/" if the user has not specified
a directory. However, creating a lambda outside of the "/"
was also causing an issue that had to be patched. In order
to allow for creating stacks outside of the root folder we
also needed to patch path and process.

Path was needed to recreate the parse method. parse is
deprecated but still used in cdk source for determining
the lock file location. Parse is called when the cwd does
not match the lambda entrypoint location.

Process was needed to allow users to set the cwd to create
entire projects in subfolders without having to always
specify absolute paths.